corrigindo conforme exemplo
