+++ semana6/LabenuForm/labenuform/src/components/Etapa2.js
@@ -15,5 +15,3 @@ class Etapa2 extends React.Component {
 }
 
 export default Etapa2;
-
-
