+++ semana6/LabenuForm/labenuform/src/components/Final.js
@@ -8,4 +8,3 @@ export default function PerguntaAberta(props) {
     </div>
   );
 }
-
