+++ semana6/LabenuForm/labenuform/src/index.js
@@ -1,12 +1,48 @@
-import { StrictMode } from "react";
+import React from "react";
 import ReactDOM from "react-dom";
 
-import App from "./App";
+import "./styles.css";
+import Etapa1 from "./components/Etapa1";
+import Etapa2 from "./components/Etapa2";
+import Etapa3 from "./components/Etapa3";
+import Final from "./components/Final";
+
+class App extends React.Component {
+  state = {
+      etapa: 1
+  };
+
+  renderizaEtapa = () => {
+    switch (this.state.etapa) {
+      case 1:
+        return <Etapa1 />;
+      case 2:
+        return <Etapa2 />;
+      case 3:
+        return <Etapa3 />;
+      case 4:
+        return <Final />;
+      default:
+        return <Final />;
+    }
+  };
+
+  proximaEtapa = () => {
+    this.setState({ etapa: this.state.etapa + 1 });
+  };
+
+  render() {
+    return (
+      <div className="App">
+        {this.renderizaEtapa()}
+        <br />
+        {this.state.etapa !== 4 && (
+          <button onClick={this.proximaEtapa}>Próxima etapa</button>
+        )}
+      </div>
+    );
+  }
+}
 
 const rootElement = document.getElementById("root");
-ReactDOM.render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-  rootElement
-);
+ReactDOM.render(<App />, rootElement);
